enhance event processing
- Company, User 인스턴스들은 여전히 무조건 DB에 저장됨
- 이는 DB의 변경과 메시지 버스의 메시지가 다르기 때문임
  - DB: CRM을 제외한 어떤 애플리케이션도 DB에 대한 접근 권한이 없다면 DB와의 통신은 식별할 수 있는 동작이 아니라 구현 세부 사항. DB의 최종 상태가 정확하기만 하면 DB 호출 횟수는 중요치 않음
  - 메시지 버스: 애플리케이션의 식별할 수 있는 동작임. 외부 시스템과의 계약을 준수하려면 CRM은 이메일 변경이 있을 때만 메시지를 메시지 버스에 넣어야 함

+++ src/main/java/com/example/refactorings/humbleobject/MessageBus.java
@@ -1,5 +1,5 @@
 package com.example.refactorings.humbleobject;
 
 interface MessageBus {
-    void sendEmailChangedMessage(int userId, String newEmail);
+    void sendEmailChangedMessage(long userId, String newEmail);
 }
